Update WAF for new bot scans.
We've started seeing an increase in invalid bot traffic for nearly a
week, notably a large number of 500s for the paths `/error` and `/Error`
from a specific IP address, and also an increase in 404s for F/CGI paths
(not used anywhere).

Also simplify the `.env*` portion of the regex while adding `.cgi`.